moving SNExTarget to another file to avoid circular import
